Fix issues related to panning to a node (#824)
When a user selects a node that is out of view, the graph explorer pans
to it and centers it. Unfortunately, this feature had some issues.

0. Panning happened before graph expansion completed. This led to pan
target calculations that relied on incorrect, outdated measurements. We
fix by waiting 250ms (the default duration for animations in the graph
explorer) before panning in order to give enough time for node expansion
to complete.

1. The graph explorer did not take into account the position of the
graph's root SVG element with respect to the viewport. This led to
panning calculations that were slightly off. We fix by taking that into
account.

2. Nodes could sometimes be occluded by the minimap - those nodes are
strictly speaking still in the view of the graph explorer, so we
previously did not pan to nodes occluded by the minimap. Now, we do.

There are still issues related to panning that are not fixed. For
instance, we do not pan to extracted nodes. Panning sometimes yields a
console error:

```js
Error: <text> attribute x: Expected length, "NaN".
```

However, this PR does fix several big issues.