Correct logic for finding if a node is in bounds
Previously, logic for determining whether a node was out of bounds in
the graph did not take into account how the top-left of the graph might
not match the top-left of the viewport.

Now, we correct that and furthermore pan to a node if it is even
partially out of view.