Make the LHS of debugger dashboard resizable (#843)
We introduce a `tf-debugger-vertical-resizer` component that lets the user resize the left hand panel by dragging the component. The component lets the user specify a min and max width (so that for instance (1) both the left and right sides remain usable and (2) the dragger component does not disappear). Logic also responds to resizing of the window.