[Projector] Distance and neighborhood function selection
Add to the Inspector panel a dropdown-menu for that distance
metric/space selection, and a second dropdown-menu for the neighborhood
function selection. Automatically update distance metric/space choices
upon change in the projection choice, and set to the newest projection
choice. Allow for future distance metric/space and neighborhood
function definitions to expand the available options.