Store left menu width in local storage and URL
We store the left menu width of the debugger dashboard within local
storage as well as within the _leftMenuWidth GET param. That way, when
the user revisits the debugger dashboard, the width will be restored.

We store the default width of the left menu within a
_DEFAULT_LEFT_MENU_WIDTH variable.

We also add 8px of padding to the main content area.