Rename left menu to left pane
We rename the left menu within the debugger plugin to the left pane.
This seems semantically more accurate. The left hand side is not really
a menu.

We store the window width as a property of the debugger dashboard so
that other methods may respond to it.

This change does not alter functionality at the user level.