Store left pane width in local storage and URL (#851)
We store the left pane width of the debugger dashboard within local
storage as well as within the _leftPaneWidth GET param. That way, when
the user revisits the debugger dashboard, the width will be restored.

We store the default width of the left menu within a
_DEFAULT_LEFT_PANE_WIDTH variable.

We also add 8px of padding to the main content area.