Rebuild graph hierarchy if callbacks changed
We make the tf-graph component rebuild the graph hierarchy if any of the handlers related to edges or nodes are set on the component. This is necessary because
the graph would render incorrectly if the graph hierarchy is built without taking into account the handlers. This fixes a race condition in the internal mapper tool in which the edge width function is occassionally not taken into consideration on certain page loads.