Let user resize tensor data and graph section
We add a dragger that lets the user resize the explore container (the section containing the graph or visualizations of tensor values) vs the tensor value overview section.

We achieve this by generalizing the `tf-debugger-vertical-resizer` component, allowing it to support both vertical and horizontal resizing.

As part of this effort, we make the height of the graph scale based on the size of its container, which solves another problem: There will no longer be a giant gap between the graph and the tensor value overview section. Previously, that gap was possible in tall viewports because the height of the graph was fixed.