Fix graph minimap by removing CSS property (#899)
Before this change, the graph explorer minimap had been broken, as
discovered by @nfelt during manual testing. Specifically, the graph
would not render within the minimap.

Surprisingly, this CSS specification within the
`tf-debugger-initial-dialog` component had been hiding the SVG element
for the rendered graph within the minimap:

```css
:host:not([_open]) {
  display: none;
}
```

That seems odd. The `:host` selector refers to the root of the shadow
DOM for the current component, in this case
`tf-debugger-initial-dialog`. That dialog is only rendered by the
debugger plugin - the debugger dashboard is not even constructed when
the graph dashboard (containing the minimap) is rendered.

I perused the vulcanized HTML and JS, and no issues seem to stand out
there.

This change solves the immediate problem of the minimap not rendering by
making the selector more specific and apply only to the component. I
will later investigate what is causing a CSS specification within the
`tf-debugger-initial-dialog` component to affect the graph explorer
minimap.