Make containers of inactive dashboards invisible
PR #902 fell short of fixing #554 because some elements such as dialogs
may be fix-positioned, in which case they anchor to the browser viewport
instead of the container. This means that may still be visible even when
the dashboard containing them is not selected.

For instance, here, we navigated to the debugger dashboard before
switching to the graph dashboard.

This fixes that problem by also making containers invisible in addition
to having a max height of 0. Note that invisible elements still have
height (unlike elements that are not displayed) - thus, Plottable can
still obtain accurate measurements of DOM elements and prevent charts
from becoming degenerate.