Correct position of context menu
Previously, the context menu was often incorrecrtly positioned due to
unexpected behavior of the layerX and layerY properties. They returned
distances in non-device spaces and thus varied based on local CSS.

Furthermore, the layerX and layerY properties will soon be deprecated.
We therefore redo the logic for positioning the context menu by making
use of the clientX and clientY properties. We subtract the offset of the
scene element from those properties. This correctly positions the
context menu.