Add /data/data_location route
This new `/data/data_location` route (of the core plugin) responds with
the location of data - either a path to a log directory or a database
URI. This new route will enable the frontend to display the data
location on the bottom left when TensorBoard runs in database mode.

I was also hoping to use this relatively innocuous change to jumpstart
discussions. For example, as we progress towards getting TensorBoard to
run in database mode, tests for plugins will have to test behavior for
both logdir and database mode. What are the norms we expect tests to
follow?

One proposal is to have tests start a logdir_based_server alongside a
db_based_server. If we later find that tests end up sharing much logic
related to starting both servers, we can refactor logic into a base
class that tests inherit from.