Fix attempt in _serve_runs() to .sort a py3 dict_keys (#1004)
Fixes #1003.  Also simplifies the code a bit overall; there's no need to pre-cache the first event times in a dict because the key function passed to sort will only be called once per element anyway (i.e. it will cache it internally).  At that point I think it's cleaner to just inline the key function to where it's used.

I also extended the testing of the run ordering to cover the no-started-timestamp cases.