Prevent content from intruding into sidebar (#1010)
Within small viewports, main content had been overlapping with the
sidebar.

![image](https://user-images.githubusercontent.com/4221553/36769398-e5c72c96-1bf8-11e8-8ce8-4b583a0df15b.png)

This change fixes that by removing a `min-width` specification on the
wrong container and placing it on the right one. Changed that min width
to 270px (from 350px). Any lower, and buttons overflow to the next line. 

New behavior:

![image](https://user-images.githubusercontent.com/4221553/36769411-eff6cb2c-1bf8-11e8-801f-16d5940d3b7d.png)