Write method to find items of active page
Previously, the image dashboard exhibited a problem in which a page of
images would duplicate when the user sets a run to off then on again.
This problem seems to stem from how polymer logic assigns a set of image
loaders to each page, and polymer somehow does not clean up a set of
image loaders for a property that had been dynamically generated (and
that we lost a reference to) such as `category._pages`.

This proposes a solution. `tf-paginated-view` generates a set of pages,
and exactly 0 or 1 of those pages are active. Instead of directly
passing `category._pages` into a dom-repeat, we remove that dom-repeat
across pages entirely (as well as the DOM-if check for whether the page
is active). We instead directly obtain the set of items from the active
page.

Test plan:

Run mnist_with_summaries. Visit the image dashboard. Hide the first
category of images. Toggle the train run several times. Re-open the
category. Note that only 1 page of images shows instead of duplicate
versions of the same page.