Reimplement video writing support in Beholder (#1046)
This change reimplements video writing in Beholder.  The new API consists of a VideoWriter class that takes a list of VideoOutput types.  When sending a frame to VideoWriter, it handles cascading through the list of outputs until it finds one that works, as well as re-initializing an output if the frame size changes, which lets us simplify logic in Beholder itself.

Right now there are only two VideoOutputs implemented.  The primary one is FFmpegVideoOutput which uses a subprocess call to FFmpeg and currently is hardcoded to generate lossless-VP9-encoded video wrapped in a .webm container at 15 FPS.  As a fallback that should work in just about any scenario, PNGVideoOutput writes individual frames as PNG files to disk using TensorFlow's own `encode_png()` op.