Make "from tensorboard.plugins.beholder import Beholder" work (#1052)
This imports Beholder and BeholderHook into the top-level `tensorboard.plugins.beholder` module so we can use that distinction to emphasize what parts of that package are public vs "internal".  I also changed the pip package BUILD rule and smoke test to include the beholder build target, so that `tensorboard.plugins.beholder` will actually be included with our pip package.