Create Beholder README.md and update docs (#1055)
Add a README.md for Beholder with some illustrative images.  It's not quite as in-depth as the original one Chris wrote, but hopefully it'll suffice for now, and the screenshots and example code are now in sync with the current code.  Also update the no-data message on the dashboard with current example code and updated links.