Make Beholder fail gracefully on non-writable config file (#1070)
This change makes Beholder fail gracefully if it can't write a config file.  Previously the plugin attempted to write a default config file to the log directory as early as `__init__()`, so any error on writing the file (e.g. a permissions error on a read-only directory) would kill TensorBoard.

With this change, Beholder only attempts to write a config file if it's already determined that it should be active (i.e. if the `Beholder` class in the model code is writing data).  This ensures that users who aren't intending to use Beholder won't have their log directories written to at all, which avoids leaving "junk" in those directories and ensures that they won't be impacted by missing write permissions.

As a further convenience, we also rewire the Beholder frontend controls to accept a disabled parameter and make the backend indicate that the controls should be disabled when it's unable to write out the config file, which should help reduce user confusion as to why the controls wouldn't be working in a case like this.