Count regular edges while picking metaedge.
The logic for patching inputs to function nodes performs a scan across
metaedges (edges that encapsulate potentially numerous base edges) in
order to determine the correct metaedge to patch. Previously, this logic
had not taken into consideration how a metaedge could represent several
regular edges. This was problematic since the index pertained to the
collective number of regular edges (not metaedges).

This had caused an internal user's graph to fail to render. The graph
contained an edge into a series node within a function. The series node
caused several regular edges to be collapsed into one metaedge.

This change has been tested with that graph as well as other graphs that
make use of TensorFlow functions. The graph WAI.