supporting streaming trace viewer in tensorboard. trace viewer normally involve huge amount of data which browser can not handle. streaming trace viewer use grpc to connect to a tpu profiler analysis service to get small part of whole trace and display to user. the amount of data depends on current view port (defined by a start and end timestamp) and resolution (defined by zoom level). currently the server reside on google owned virtual machine. there are plans to open source those part.
to avoid confusing user, the patch is done in a backward compatible
fashion. user are expected to use old trace viewer (which limit the
event to 1 million), if the user specify the MASTER_TPU environment to
the IP address of the green VM, we will try to estabilish a grpc stub to
that IP:8466 port. only if stub is successfully connected we will use
new tools 'trace_viewer@' to OVERRIDE the old trace_viewer. Override
will hide old trace viewer tools to avoid confusing user.

Upon receiving the tool/data probing in python, we will determine which
tools to use. using 'trace_viewer@' will signal the downstream
componenet that streaming trace viewer is in use.

Most of the trace viewer related code already exist within google3. I
removed host filters etc for the reason that we don't support multiple
hosts yet.