Let developer set rank direction for graph
We add a `rankDirection` property to the `tf-graph-loader` component.
This property is passed to the dagre layout engine as the 'rankdir'
graph option. We hence let the user change which way the graph flows.
For instance, we can let the graph flow from top to bottom.

This PR is a small part of generalizing the graph explorer.