Find runs by glob-ing levels of subdirectories (#1087)
Previously, TensorBoard found runs within the log directory by
performing a `tf.gfile.Walk` across directories. That turned out to be
very slow for users with expansive log directories (with many
subdirectories and files) because `tf.gfile.Walk` makes many calls at
the python level to C++ logic for reading data from disk (as opposed to
performing much of that logic directly at the C++ level). Disk reads can
be very costly for certain (such as remote and distributed) file systems.

We find runs (subdirectories) much more quickly by iteratively glob-ing
deeper and deeper directory levels. For a certain internal user, this
reduced the time it took to find all runs from 20 minutes to 20 seconds.

We already have tests that check whether the multiplexer correctly finds
subdirectories that are runs, so this pull request adds no further
tests. We preserve that correctness and just speed up that logic. This
change may result in us finding runs in different orders, but that is
fine - TensorBoard documents no guarantees in terms of the order in
which runs are discovered.

Thank you to @nfelt for this idea and the fruitful brainstorming
session.