Make debugger frontend gracefully handle backend errors
The backend of the debugger plugin may error now and then and respond with a status 500. In those cases, the frontend should not leave the user in an invalid state such as one in which the debugger dashboard indicates a false state of continuing.

Furthermore, we prevent the frontend from selecting nodes that do not exist, which may cause a JavaScript error. Otherwise, the debugger dashboard might try to select a non-existent node if the backend errs.