Let user search nodes by regex
This change introduces a search box to the graph explorer that lets the user input a regex and then performs a partial regex match on all nodes (both metanodes and ops) within the graph.

When the user clicks on a match, the graph selects the node and pans to it if it is out of view.