Remove a couple un-needed lines
We remove the statement that loads the
`tensorboard_webcomponent_library` symbol into the BUILD
file for the `tf_graph_node_search` component.
`tensorboard_webcomponent_library` is not used.

We remove an empty line from the BUILD file for the
`tf_graph_node_search` component.

We remove an un-needed dep from the tf_graph_controls BUILD
target.