Disable Pylint
Pylint is a nice tool, but hasn't been that helpful for TensorBoard.

- See https://travis-ci.org/tensorflow/tensorboard/jobs/376968526 where it
  takes more than two minutes to run, representing >35% of Travis CI latency.

- We typically have to wait until the very end of a Travis run before we get
  Pylint results, because we run it in a really hackish way. This is also
  likely the cause of its poor performance and has resulted in surprising
  breakages, such as the one we encountered in #1206.

- Pylint has a very different point of view on style / best practices compared
  with Google and IntelliJ.

- TensorBoard doesn't work with the latest version of Pylint.

IntelliJ is good enough, in my personal opinion. I'd be open to keeping Pylint
if it gave quick feedback as part of the normal Bazel build process. I looked
into the difficulty of doing that and it was discouraging.