Fixed the dangling interaction issue (#1287)
Interactions do not get destroyed or deactivated automatically
even if a component it is attached to is detached from DOM. As
a result, event listeners were eagerly listening to mouse events
and expected a component to be mounted.

Cleaned up vz-line-chart a bit by factoring interaction out of
tooltip rendering. Interaction is supposed to be reused over
time.