Refactoed: pass datum to chart-data-loader
The loader, previously, was very geared towards runs and tag
combination which can be generalized to any datum. The loader does
not have to be aware of the application logic like dataUrl or
data parsing logic that it can be more generalized to take in
arbitrary piece of data.

While refactoring, the change pulled out the loading logic into
a reusable behavior.