Pipe the tags information to selection
Background: Tags are created for a specific plugin.
Reason: Categorization is used near visual component to distill
tagfilter/tag/experiment/runs grouping from a selection. It is also
important to note that a categorization will wildly differ depending
on which plugin is invoking. To be able to filter out unrelated
tag categorization, now, categorization utils take the pluginName
to filter out unrelated tags based on information we pipe from the
backend.