Let users blacklist subdirectories by regex
Introduce a subdirectory_blacklist_regex flag to TensorBoard. If
provided, TensorBoard excludes data from any subdirectory with an
absolute path that partially matches this regular expression.

This flag allows users to point TensorBoard to a path containing many,
many subdirectories and exclude runs that are not interesting for
comparison. Users could toggle runs in the UI, but that requires effort
when there are many runs/experiments. Subdirectories that are filtered
away by the regular expression are logged for later debugging.