Use QueryParam for data selector config
URL storage uses two mediums when bootstrapping -- before the TB
application is up, it uses fake hash, then it uses the real URL hash.
This is done for test predictability.

Since whether to use data selection or not is related to the
experiment control and not storage of some user data, we decided to use
the query parameter instaed.