Fixed the zombie selection bug (#1382)
When removing an experiment, because of debounce and asyncness,
selection that has been removed gets reinserted to the list of
selection. This is fixed by removing the debounce and making the higher
order component, data-selector, emit-less change when data does not
change. Also, it added some safeguard where we ignore events from a
removed experiments.

Also, we now populate the internal state of `_dataReady` correctly.