FIX: Recolor when color scale changes for run
When list of runs changes, runsScale remaps runName to color. The
change has to be reflected by UI but tooltip or run selector dropdown
reflected the color change.

The change addresses the bug described above but the change is not
final -- a runsColorScale change is picked up by components that also
listens to change of the runsStore, but this is by luck and not design.
RunsColorScale which currently is a function of runs in the runsStore
should emit change separately and componennts should not be aware of
this dependency when consuming the color. The fix does not address this
issue and tf-data-select-row currently has an issue because it does not
listen to the runsStore.