BUG: make select-row update when shifted (#1387)
Say there were two selection: A and B. When removing A, Polymer does not
remove DOM for A but, rather, one for B. As a result, A will be updated
with parameter of B and it will need to be reinitialized including
binding to the storage.

When updated, not only should it refetch the list of runs, but also needs
to detach event listeners with the previous storage keys. If not done,
any change in hash will trigger the callback and set the value
incorrectly.