REFACTOR: make series name separate from title (#1383)
Series displayName needs to update based on whether user is comparing
multiple experiment or not. However, changing the name, which was previously
used as an id to LineChart class, caused various issues. 

We solved it by persisting an arbitrary piece of data in the metadata (of
Plottable.Dataset) and use custom TooltipColumns to create displayName
for a series.