FIX: line-chart is rendered properly (no-shrink)
When using paginated to scroll through charts quickly, you will trigger
the chart creation (which is debounced for 350ms). The chart creation,
even if it is no longer in foreground and is "display:none"d, it will
resume drawing and result in very shrunken line-chart. There has been
some detector of such bad draw but it did not account for the debounce
logic in the line-chart.

To mitigate the problem, we now listen to on Component#onAnchor event to
listen to DOM mounting to detect whether we have a bad draw: drawn in the
background.

Alternatively, we can cancel debounce when tf component detects the
backgroundness by either creating a new state in chart, `active`, or
a method `cancelMakeChart`.