FIX: line-chart is rendered properly (no-shrink) (#1403)
When using paginated to scroll through charts quickly, you will trigger
the chart creation (which is debounced for 350ms). The chart creation,
even if it is no longer in foreground ("display:none"d), it will draw and result
in a shrunken line-chart. There is a bad draw detection logic to remedy the
problem, but it did not account for the debounce logic in the line-chart.

To mitigate the problem, we now listen to on Plottable.Component#onAnchor
event to known when a chart is mounted on DOM and know when a drawn
happened while it was in the background.

Alternatively, we can cancel the chart creation debounce by either creating
a new state in chart, active, or a method cancelMakeChart.