REFACTOR: mv tooltip container out of scroll area (#1406)
Chart tooltip (or all popovers) should render at the body level in order
to prevent clipping from a scrollable area. To be more specific, even
position:fixed can get contained inside a region if it is within a
scrollable region. As a remedy, most tools render popover in children of
body.

To achieve this in a component structure, there are two competing goals:
1. intuitive component structure (tooltip is a subcomponent chart)
2. DOM-wise, tooltip should not be nested in chart but should be at body
   level.
Libraries like React allows such pattern via "tunneling". By manually creating
an element in the body, we emulated the behavior.

In this commit, we:
- introduced vz-chart-tooltip that knows how to position
- applied the new tooltip to scalar-card
- introduced new positioning, "auto"