FIX: custom scalar now reacts to run selection
With tf-line-chart-data-loader refactor which now takes `dataToLoad`
separately from `dataSeries` (latter defines series that are to be drawn
onto the chart whereas the former defines data requirement), the custom
scalar plugin broke because it failed to take run selection when setting
the `dataSeries` property.

The change also addresses the issue where smoothed series do not show in
fade.