Fixed some shrunk histogram issue
Repro step
---
1. launch with an empty tag regex
2. type a character (that'd result in few matches) then delete it
   quickly
3. after few moment, type the character again

Cause
---
dom-if has issue where if="false" does not remove the DOM but rather
simply hides the slot.
https://github.com/Polymer/polymer/blob/1.x/src/lib/template/dom-if.html#L127

While the pane is briefly shown in step #2, histogram-loader fetches
data and before it has chance to draw the series, dom-if can hide the
DOM. As data comes back, histogram tries to draw based on the bounding
box that is shrunken to 0x0 since its container is no longer displayed.
To work around these issues, we now listen to category-pane show event
to redraw all histogram.