Created modes to data selector
For most of the user, data-selector-alt is the appropriate UI while
data-selector is more appropriate for more advanced user who wants to
fine tune what to compare.

Instead of TB deciding what to display, we now make use of both of them
with a simple toggler to toggle between two modes.