Fix observers when using data selector
Polymer's complex observer require values it listens to be instantiated.
`tagFilter` was previously removed because tf-tag-filterer took the
value from its parent instead of bootstrapping from the storage.

This change ensures that complex observers are properly set up with all
listed properties instantiated and to take all consumed properties in
callback to be listed in the observer definition.