Made data-selector dropdown react to keyboard
For many dropdown components in the TF, we were nesting an input to an
effective button. This seems to react weirdly to keyboard events in
general and would not respond properly to ENTER or SPACE to activate a
focused button.

We created an `paper-input-container` like component that aesthetically
look similar that would not have the same problem as the input.