Fixed tagFilterer to correctly bootstrap and work
tf-tag-filterer, correctly, defined its property `tag-filter` as
readOnly and this broke the data connection between `paper-input`.
Between the paper-input and the filterer, there should be a
two-way binding of its value whereas all the consumers of the
tf-tag-filterer should only read from the component.

Now, using the private property, `_tagFilter`, we establish the two-way
binding while the public property is only used to publish the change.

Note: bug was introduced over #1483, #1475, and #1493.