Downsample histogram tensor before conversion
In one of the realistic event file, we had
N=len(tensor_events)=500, M=max(len(event.tensor_proto))=110,000.
When making burst of requests from the histogram plugin, this large
c*N*M the /histograms endpoint to perform really poorly (c=number
of histogram charts on the screen). tf.make_ndarray ideally should
perform better wrt to the input but it can be done/studied separately.

Another way to mitigate the issue is to use ForkingWSGIServer instead
of ThreadedWSGIServer as, in this case, compute power is the bottleneck.
However, we do not have too many compute intensive data endpoints and it
is rather an overkill.