Fixed graph edge size computation
When we upgraded the Lodash from 3 to 4, it broke one of the private
method used to comput the size. According to
https://lodash.com/docs/4.17.10#lodash, `sum` is no longer chainable
breaking the method.

Though we tried to add tests to detect such breakage but graph code
is massive nad is largely untested. Since we cannot add all the tests at
once, we added one simple test, as a start.