Fix the issue where minimap content gets clipped
Bug: https://github.com/tensorflow/tensorboard/issues/1574
Cause: the graph contains no main graph node and all nodes are
auxilliary nodes. As a result, the drawn nodes are shifted right by
predefined amount for visuals. When measuring the bounding box of the
scene to draw the minimap, we were measuring using
SVGGraphicsElement.getBBox which returned BBox without any whitespace.
As a result, although we wanted the width from the origin to right edge
of the right most node, we were omitting the origin to left most edge
of the content resulting in clipped content in the minimap.