Fix the issue where minimap content gets clipped (#1600)
Bug: #1574

Cause: the graph in the bug contains no main graph node and all nodes
are auxiliary nodes. As a result, the drawn nodes are shifted right by
predefined amount for visuals.

When measuring the bounding box of the scene to draw the minimap,
we were using SVGGraphicsElement.getBBox which returned a tight BBox
without concerning any whitespace. As a result, although we wanted the
width from the origin to right edge of the right most node, we were omitting
the origin to left most edge, resulting in clipped content in the minimap.

Fix: measure the scene width from origin to the right edge of the content.