docs: fix vz_line_chart smoothing documentation (#1622)
Summary:
Fixes #1621. These docs were far too closely specified. There was a day
where they also happened to be correct, but it is not this day.

Test Plan:
Look at the implementation of `resmoothDataset` in both versions of the
code (v1: `vz-line-chart.ts`, v2: `line-chart.ts`; the implementations
are the same). Observe that the implementation is not a moving average
whose window size is set by the `smoothingWeight`; instead, it is the
attenuation parameter of a low-pass filter.

wchargin-branch: docs-smoothing