Infer name of exp when using db_import (#1616)
Previously, user had to supply name of the experiment when passing
logdir. While this guaranteed correctness[1], it caused some friction in
adopting the db_import so, when a name is not passed, we infer the name
from the first top level in subdirectories of logdir.

Note that this inference is crude and will lead to some proliferation of "."
in experiment names as well as run names.

[1]: In case user places event file in say:
- /logdir/event.file
- /logdir/eval/event.file
the experiment name and run name will be screwed up.