docs: add instructions for building old revisions
Summary:
When bisecting or exploring the history, it can be useful to build old
commits. Unfortunately, due to backward-incompatible changes in Bazel,
this is not possible natively. This commit describes a workaround.

It’s unfortunately not too discoverable in `DEVELOPMENT.md`, but at
least it’s somewhere.

Test Plan:

    $ git checkout -q "$(git rev-list origin/master --before=2018-08-01 -n 1)"
    $ bazel build //tensorboard 2>&1 | tail -n 1
    FAILED: Build did NOT complete successfully
    $ git cherry-pick bc4e7a6e5517daf918433a8f5983fc6bd239358f >/dev/null
    $ bazel build //tensorboard 2>&1 | tail -n 1
    INFO: Build completed successfully, 117 total actions

wchargin-branch: docs-build-old