test: add `tf_web_test` Starlark macro
Summary:
This macro removes the need to create a `test.py` target for each web
test, and centralizes some common options like browser selection,
default test timeout, and flakiness.

A benefit of this change is that it removes the direct dependency on the
webtesting rules in each BUILD file, instead centralizing the dependency
in `web.bzl`. This makes it easier to change this dependency when we
import into Google (where the infrastructure is different).

Test Plan:

  - Run `bazel test //tensorboard/components/vz_sorting/test`, which
    should pass.
  - Change `sortingTests.ts` to make a test fail. Re-run `bazel test`,
    which should fail after 3 tries.
  - Add `flaky = False` to the `vz_sorting/test:test` target. Re-run
    `bazel test`, and note that it fails immediately.
  - Run with `bazel run` instead of `bazel test` and note that the test
    runs in a browser (not headless).

wchargin-branch: tf-web-test